requirements.txt
requests
boto3
python-dotenv
streamlit
pandas 
plotly
numpy